feat: v0.183.0 -- svr_implemented_options fixed to Table 18's real 8-bit layout, REQ-RMAP-004..008 retired (REQ-RMAP-030, issue #200)
TC18 §12.7.5 Table 18 defines svr_implemented_options as an 8-bit
read-only register at 0x0016 with one independent bit per optional
feature (a=compound&wait, b=trigger, c=chained, d=time synch/timed,
e=enhanced cancellation). c-RCP instead declared a 32-bit field
carrying six bits of this project's own invention, grouped into three
all-or-nothing pairs and enforced by
rcp_regmap_options_group_consistent() -- with no bit at all for
trigger or chained requests, even though c-RCP implements both.

This directly contradicted REQ-RMAP-004..008, which had already
closed by describing that same invented pairing design, citing
§12.9.1.1 as their basis. Neither side could self-verify without the
real specification text, so this was surfaced to the user rather than
resolved by guessing; the user chose to pause and verify against the
actual TC18 specification.

That verification located OA_TC18_specification_v_0.5.1_RC.pdf, the
real 117-page TC18 spec (previously unknown to exist in this project's
search history -- prior searches only ever checked for .txt files).
Reading it directly: Table 18 confirms REQ-RMAP-030's reading exactly
-- five independent bits, no pairing concept. §12.9.1.1,
REQ-RMAP-004..008's shared citation, is titled "Handling multiple
requests in incoming messages" and is entirely about an RC Server
processing several ACF-type requests packed into one AVTPDU frame; it
says nothing about svr_implemented_options, feature advertisement, or
any bit-pairing rule. The exact phrase these five requirements quote
does not appear anywhere in that section.

REQ-RMAP-004..008's citation is a genuine misattribution. Their
pairing rule has no TC18 basis and is retired outright (not
deprecated-then-removed -- there was no correct behavior to preserve a
transition window for).

Production changes:
- rcp_regmap_general_t.svr_implemented_options: uint32_t -> uint8_t
- Six invented, paired RCP_REGMAP_OPT_* constants replaced by five
  independent single-bit constants matching Table 18's bit order
- rcp_regmap_options_group_consistent() removed outright
- rcp_timed_feature_enabled(): uint32_t -> uint8_t, single-bit check
- cli.c's capabilities_json() "features" array: gained "trigger" and
  "chained", closing a real, previously-unadvertised capability gap
- config.c's manifest parser: gained matching "trigger"/"chained"
  names in OPTION_BIT_NAMES

Test changes:
- test_regmap.c: 4 group-consistency tests replaced with 1 proving
  independent bit settability
- test_tc18_gaps_regmap.c: deviation pin rewritten positive
- test_request_timed.c: simplified to single-bit checks
- test_config.c: new test proving "trigger"/"chained" parse correctly
- test_cli.c: new test inspecting capabilities' "features" array
  content -- no prior test in that file did this at all

.fusa-reqs.json: REQ-RMAP-030 stays partial/tc18-gap (0x0016 is still
past the discovery slice's 0x000D wire-reachability ceiling,
REQ-RMAP-024, still open) with fully rewritten text. REQ-RMAP-004..008
each gain a new "retired" status/scope and fully rewritten text
explaining the primary-source finding.

Mutation-tested five ways: three isolated-logic mutations (timed
feature bit-check inversion, CLI trigger-bit zeroing, config
trigger-name bit swap), each caught by its own test; one full
seven-file production revert with every touched test file's changes
kept, breaking the build across four test files and confirming real
dependency on the fix, not just updated expectations. All mutations
restored clean, diff-verified byte-identical against pre-mutation
backups.

Full suite (65/65 ctest suites) + ASan/UBSan clean. Fresh cfusa check
(0 errors) + all three separate cfusa trace invocations (100%
req-coverage, 100% sec-tested, 0 untested).

+++ include/rcp/request_timed.h
@@ -91,8 +91,8 @@
  *
  * rcp_timed_feature_enabled() reads (but the register map module itself
  * is not modified by this file) regmap.h's already-published
- * RCP_REGMAP_OPT_TIME_SYNC_TSCF/_PRESENTATION bits -- both must be set for
- * a server to accept Timed requests at all, since a Timed request is
+ * RCP_REGMAP_OPT_TIME_SYNC bit (REQ-RMAP-030) -- it must be set for a
+ * server to accept Timed requests at all, since a Timed request is
  * meaningless without the same underlying time-sync capability TSCF
  * itself depends on.
  */
@@ -144,10 +144,11 @@ const char *rcp_timed_strerror(rcp_timed_errc_t e);
 
 /* ── Feature gating ─────────────────────────────────────────────────────────── */
 
-/* True iff both RCP_REGMAP_OPT_TIME_SYNC_TSCF and
- * RCP_REGMAP_OPT_TIME_SYNC_PRESENTATION are set in options -- see the
- * file header. */
-bool rcp_timed_feature_enabled(uint32_t options);
+/* True iff RCP_REGMAP_OPT_TIME_SYNC (regmap.h, REQ-RMAP-030's own
+ * single "d: time synch and timed requests" bit -- retyped from this
+ * function's own former two-bit-pair check, REQ-RMAP-030) is set in
+ * options -- see the file header. */
+bool rcp_timed_feature_enabled(uint8_t options);
 
 /* ── Timed request encode/decode ──────────────────────────────────────────── */
